refactor: 모달 디렉토리 분리
